Testing mobile features, DOES NOT COMPLETELY FUNCTION
mobile detection for screen optimization, and ability to navigate through games on mobile, as well as adding some assets, menu redesign, and ability to select to fight computer (currently broken)

+++ States/js/css.js
@@ -15,6 +15,13 @@ var cssState={
     game.physics.arcade.enable(chickIcon);
     chickIcon.tint =  0xffffff;
 
+    //TEST:COMPUTER icon
+    computerIcon = game.add.sprite(game.world.width * .5, game.world.height * .5  - 100, 'computerIcon');
+    computerIcon.anchor.setTo(.5,.5);
+    computerIcon.scale.setTo(.25,.25);
+    game.physics.arcade.enable(computerIcon);
+    computerIcon.tint =  0xffffff;
+
     player1Icon = game.add.sprite(game.world.width * .5 -50, game.world.height * .5, 'player1cssIcon');
     player2Icon = game.add.sprite(game.world.width * .5 +50, game.world.height * .5, 'player2cssIcon');
 
@@ -37,6 +44,9 @@ var cssState={
 
     dudeIcon.enableBody = true;
     chickIcon.enableBody = true;
+
+    buttonSound = game.add.audio('buttonSound');
+
     var startLabel=game.add.text(80,game.world.height-40,'Press "1" key to play game after selecting characters!',{font: '25px Arial',fill:'#ffffff'});
     //var wkey= game.input.keyboard.addKey(Phaser.Keyboard.W);
     //wkey.onDown.addOnce(this.start,this);
@@ -46,6 +56,17 @@ var cssState={
     player1BodyIcon = game.add.sprite(game.world.width * .25, game.world.height * .75, '');
     player2BodyIcon = game.add.sprite(game.world.width * .75, game.world.height * .75, '');
 
+
+
+
+
+    if(game.device.android || game.device.iOS) //If on Mobile, make an invisible button that will be visible only when the characters are selected
+    {
+      startButton = game.add.button(game.world.width *.25,game.world.height * - 100, 'startButton');
+      startButton.visible = false;
+      startButton.onInputUp.add(this.start,this);
+    }
+
 //TODO:Incorperate dragUpdate function event system into current system. I think it's needed to fix bugs/add dynamic features like spawning the character when hovering over while still dragging.
 //TODO:
 //find a way to change text, show sprite and name with alpha applied when hovering but NOT selecting character, SOLUTION: probably above comment
@@ -53,7 +74,7 @@ var cssState={
   },
   start: function(){
     music.stop();
-    
+
    game.state.start('play');
  },
  update: function() {
@@ -76,30 +97,31 @@ var cssState={
      //Eventually allow the player to start game;
      gameReadyText.text = `Game ready`;
      game.state.start('play');
-
-
-
-
    }
    else if(charSelected1 && charSelected2)
    {
      //Eventually allow the player to start game;
      gameReadyText.text = `Game ready`;
-     game.state.start('play');//temporary fix, edit later
-
+     if(game.device.android || game.device.iOS)
+     {
+       startButton.visible = true; //If on mobile and the characters are selected, make the start button visible
+     }
+    // game.state.start('play');//temporary fix, edit later
    }
    else {
      {
        gameReadyText.text = ``;
+       startButton.visible = false; //Hide the start button and text that says game is ready when both players have not selected a character
      }
+
    }
  },
  onDragStop: function() {
 
-//If you drop the curser on the icon
+//If you drop the cursor on the icon
    if(game.physics.arcade.overlap(player1Icon, dudeIcon))
    {
-
+     buttonSound.play();
      //Determine's what's spawned, and lets you start game
      charName1 = "dude";
      charSelected1 = true;
@@ -111,7 +133,7 @@ var cssState={
      player1BodyIcon = game.add.sprite(game.world.width * .25 - 100, game.world.height * .5, 'dude');
 
      player1BodyIcon.scale.setTo(3.5,3.5);
-     player1BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player1BodyIcon.animations.add('idle', [1, 2], 5, true);
      player1BodyIcon.animations.add('kick', [6], 5, true);
      if(player1BodyIcon.animations)
      {
@@ -126,6 +148,7 @@ var cssState={
    //If you drop the icon on the chick Picture
    if(game.physics.arcade.overlap(player1Icon, chickIcon))
    {
+     buttonSound.play();
      charName1 = "chick";
      charSelected1 = true;
      chickIcon.tint =  0xffff00;
@@ -134,7 +157,7 @@ var cssState={
      player1BodyIcon = game.add.sprite(game.world.width * .25 - 100, game.world.height * .5, 'chick');
 
      player1BodyIcon.scale.setTo(3.5,3.5);
-     player1BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player1BodyIcon.animations.add('idle', [1, 2], 5, true);
      player1BodyIcon.animations.add('kick', [6], 5, true);
      if(player1BodyIcon.animations)
      {
@@ -146,6 +169,7 @@ var cssState={
 
    if(game.physics.arcade.overlap(player2Icon,dudeIcon))
    {
+     buttonSound.play();
      charName2 = "dude";
      charSelected2 = true;
      dudeIcon.tint =  0xffff00;
@@ -153,7 +177,7 @@ var cssState={
 
      player2BodyIcon = game.add.sprite(game.world.width * .75 - 100, game.world.height * .5, 'dude');
      player2BodyIcon.scale.setTo(3.5,3.5);
-     player2BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player2BodyIcon.animations.add('idle', [1, 2], 5, true);
      player2BodyIcon.animations.add('kick', [6], 5, true);
      player2BodyIcon.visible = true;
 
@@ -172,6 +196,7 @@ var cssState={
 
    if(game.physics.arcade.overlap(player2Icon,chickIcon))
    {
+     buttonSound.play();
      charName2 = "chick";
      charSelected2 = true;
      chickIcon.tint =  0xffff00;
@@ -179,7 +204,31 @@ var cssState={
 
      player2BodyIcon = game.add.sprite(game.world.width * .75 - 100, game.world.height * .5, 'chick');
      player2BodyIcon.scale.setTo(3.5,3.5);
-     player2BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player2BodyIcon.animations.add('idle', [1, 2], 5, true);
+     player2BodyIcon.animations.add('kick', [6], 5, true);
+
+     if(player2BodyIcon.animations)
+     {
+       player2BodyIcon.alpha = 1;
+     }
+   }
+   else
+   {
+    // player2BodyIcon.kill();
+   }
+
+   if(game.physics.arcade.overlap(player2Icon,computerIcon))
+   {
+     buttonSound.play();
+     charName2 = "chick";
+     charSelected2 = true;
+     computerIcon.tint =  0xffff00;
+     player2BodyIcon.kill();
+     controlOptionAI = -2; //Temporary till we have the AI logic, then replace this with a -2 instead,using vpad to test functionality
+     console.log("controlOptionAI: " + controlOptionAI);
+     player2BodyIcon = game.add.sprite(game.world.width * .75 - 100, game.world.height * .5, 'chick');
+     player2BodyIcon.scale.setTo(3.5,3.5);
+     player2BodyIcon.animations.add('idle', [1, 2], 5, true);
      player2BodyIcon.animations.add('kick', [6], 5, true);
 
      if(player2BodyIcon.animations)
@@ -197,7 +246,7 @@ var cssState={
      player1BodyIcon.kill();
    }
 
-   if(!game.physics.arcade.overlap(player2Icon,dudeIcon) && !game.physics.arcade.overlap(player2Icon,chickIcon))
+   if(!game.physics.arcade.overlap(player2Icon,dudeIcon) && !game.physics.arcade.overlap(player2Icon,chickIcon) && !game.physics.arcade.overlap(player2Icon,computerIcon))
    {
      player2BodyIcon.kill();
    }
@@ -250,6 +299,19 @@ var cssState={
      charName2 = "";
      charSelected2 = false;
      chickIcon.tint =  0xffffff;
+
+      if(player2BodyIcon.animations)
+      {
+        player2BodyIcon.alpha = .5;
+      }
+   }
+   if(game.physics.arcade.overlap(player2Icon, computerIcon))
+   {
+     charName2 = "";
+     charSelected2 = false;
+     computerIcon.tint =  0xffffff;
+     controlOptionAI = 2;
+     console.log("controlOptionAI: " + controlOptionAI);
 
       if(player2BodyIcon.animations)
       {
